+++ middleware/src/types/index.ts
@@ -1,3 +1,3 @@
-export * from "./nodeStatusResponse";
+export * from "./node-status-response";
 export * as OnChain from "./on-chain";
 export type Sort = "DESC" | "ASC";
