+++ middleware/src/types/node-status-response.ts
